fix: Balance calculations
