+++ lib/core/database/sql/queries/select-full-data.drift
@@ -55,6 +55,13 @@ getLatestValuationForAccount(:accountId AS TEXT):
     ORDER BY date DESC
     LIMIT 1;
 
+getLatestValuationForAccountAtDate(:accountId AS TEXT, :date AS DATETIME):
+    SELECT *
+    FROM valuations
+    WHERE accountId = :accountId AND date <= :date
+    ORDER BY date DESC
+    LIMIT 1;
+
 getValuationsForAsset(:assetId AS TEXT):
     SELECT *
     FROM valuations
