+++ lib/core/presentation/widgets/time_series_evolution_chart.dart
@@ -82,7 +82,8 @@ class TimeSeriesEvolutionChart<T> extends StatelessWidget {
                         spot.x.toInt(),
                       );
 
-                      final text = '${DateFormat.yMMMd().format(date)}\n';
+                      final text =
+                          '${DateFormat.yMMMd().format(date)}\n${spot.y.toStringAsFixed(2)}';
 
                       return LineTooltipItem(
                         text,
